fix batch job mappings
